Added UpdatePassword, Active and Admin
I added the ability to activate/deactivate a user. Added admin tags to each user (off by default, I will have to manually create the admin users later).

There is also a new table UpdatePassword which has requests to update the password by a user, which are then deleted by the deleteUpdatePassword when they are 24 hours or older (this is not a perfect solution, as it runs every 24 hours, so some requests will last longer than 24 hours.)